Don't toggle read status on left-click unless the click mode is New Window.
MangaDex is throwing errors if we try to toggle the read status while a page navigation is taking place.  This change makes it so we only toggle the read status on a left-click if we enable the option to open left clicks in a new window.  There is no need to flip the read status while we are navigating away from the page in the first place.